refactor: Remove UBYouTubePublisher
Publishing to YouTube was non-functional due to API changes on their part.
Since noone reported issues with it, it is reasonable to assume that there
are no users. Since the functionality is unmaintained, remove it entirely.

In previous commits, the upload function was made inaccessible. Now remove
the actual files.